a bit more refactoring for speperation of concerns xd
