+++ frontend/lib/ui/screens/login_screen.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import '../../providers/user_provider.dart';
 import '../widgets/login_widget.dart';
 
 class LoginScreen extends StatelessWidget {
