+++ frontend/lib/ui/widgets/gamepad_widget.dart
@@ -20,17 +20,7 @@ class _GamepadWidgetState extends State<GamepadWidget> {
             GestureDetector(
               onTapDown: (_) => carControlProvider.sendCommand('car/control', '1'),
               onTapUp: (_) => carControlProvider.sendCommand('car/control', '0'),
-              child: AnimatedContainer(
-                duration: Duration(milliseconds: 200),
-                height: constraints.maxWidth > 400 ? 80 : 60,
-                width: constraints.maxWidth > 400 ? 80 : 60,
-                decoration: BoxDecoration(
-                  color: Colors.blue,
-                  borderRadius: BorderRadius.circular(30),
-                  boxShadow: [BoxShadow(color: Colors.black26, blurRadius: 10)],
-                ),
-                child: Icon(Icons.arrow_upward, size: 40, color: Colors.white),
-              ),
+              child: _buildButton(constraints, Icons.arrow_upward),
             ),
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
@@ -38,54 +28,38 @@ class _GamepadWidgetState extends State<GamepadWidget> {
                 GestureDetector(
                   onTapDown: (_) => carControlProvider.sendCommand('car/control', '6'),
                   onTapUp: (_) => carControlProvider.sendCommand('car/control', '0'),
-                  child: AnimatedContainer(
-                    duration: Duration(milliseconds: 200),
-                    height: constraints.maxWidth > 400 ? 80 : 60,
-                    width: constraints.maxWidth > 400 ? 80 : 60,
-                    decoration: BoxDecoration(
-                      color: Colors.blue,
-                      borderRadius: BorderRadius.circular(30),
-                      boxShadow: [BoxShadow(color: Colors.black26, blurRadius: 10)],
-                    ),
-                    child: Icon(Icons.arrow_back, size: 40, color: Colors.white),
-                  ),
+                  child: _buildButton(constraints, Icons.arrow_back),
                 ),
                 SizedBox(width: 20),
                 GestureDetector(
                   onTapDown: (_) => carControlProvider.sendCommand('car/control', '5'),
                   onTapUp: (_) => carControlProvider.sendCommand('car/control', '0'),
-                  child: AnimatedContainer(
-                    duration: Duration(milliseconds: 200),
-                    height: constraints.maxWidth > 400 ? 80 : 60,
-                    width: constraints.maxWidth > 400 ? 80 : 60,
-                    decoration: BoxDecoration(
-                      color: Colors.blue,
-                      borderRadius: BorderRadius.circular(30),
-                      boxShadow: [BoxShadow(color: Colors.black26, blurRadius: 10)],
-                    ),
-                    child: Icon(Icons.arrow_forward, size: 40, color: Colors.white),
-                  ),
+                  child: _buildButton(constraints, Icons.arrow_forward),
                 ),
               ],
             ),
             GestureDetector(
               onTapDown: (_) => carControlProvider.sendCommand('car/control', '2'),
               onTapUp: (_) => carControlProvider.sendCommand('car/control', '0'),
-              child: AnimatedContainer(
-                duration: Duration(milliseconds: 200),
-                height: constraints.maxWidth > 400 ? 80 : 60,
-                width: constraints.maxWidth > 400 ? 80 : 60,
-                decoration: BoxDecoration(
-                  color: Colors.blue,
-                  borderRadius: BorderRadius.circular(30),
-                  boxShadow: [BoxShadow(color: Colors.black26, blurRadius: 10)],
-                ),
-                child: Icon(Icons.arrow_downward, size: 40, color: Colors.white),
-              ),
+              child: _buildButton(constraints, Icons.arrow_downward),
             ),
           ],
         );
       },
+    );
+  }
+
+  Widget _buildButton(BoxConstraints constraints, IconData icon) {
+    return AnimatedContainer(
+      duration: Duration(milliseconds: 200),
+      height: constraints.maxWidth > 400 ? 80 : 60,
+      width: constraints.maxWidth > 400 ? 80 : 60,
+      decoration: BoxDecoration(
+        color: Colors.blue,
+        borderRadius: BorderRadius.circular(30),
+        boxShadow: [BoxShadow(color: Colors.black26, blurRadius: 10)],
+      ),
+      child: Icon(icon, size: 40, color: Colors.white),
     );
   }
 }
