+++ frontend/lib/ui/widgets/login_widget.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import '../../providers/user_provider.dart';
 import '../../providers/car_control_provider.dart';
 import 'package:flutter_animate/flutter_animate.dart';
