+++ frontend/lib/utils/constants.dart
@@ -1,5 +1,5 @@
 const WEBSOCKET_URL = 'ws://192.168.0.165:8181';
 const double kMobileWidth = 320.0;
 const double kMobileHeight = 240.0;
-const double kWebWidth = 800.0;
-const double kWebHeight = 600.0;
+const double kWebWidth = 320.0;
+const double kWebHeight = 240.0;
